requirements.txt
# Data
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0

# ML/DL
torch>=2.0.0
pytorch-lightning>=2.0.0

# Time Series
statsmodels>=0.14.0
statsforecast>=1.6.0

# Optimization
scipy>=1.10.0
cvxpy>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
